Health chest added, Healthbars fixed, Animated sprites added
- Fixed enemy healthbars
- Fixed boss healthbars
- Fixed TakeDamage() to take int values instead of float
- Applied scripted animations to both boss and sheep projectiles
- Sheep projectiles now bounce

+++ Assets/Scripts/EnemyAI/EnemyAI.cs
@@ -64,7 +64,7 @@ public class EnemyAI : MonoBehaviour
         {
             health -= bullet.dmg;
             if (healthBar)
-                healthBar.transform.localScale = new Vector2(hbScale.x * (health / maxHealth), hbScale.y);
+                healthBar.transform.localScale = new Vector2(hbScale.x * ((float)health / (float)maxHealth), hbScale.y);
         }
     }
 }
